+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changes
 * Removed gfdatasource - feature is built in to Grafana since v5.
 * Generate API docs for readthedocs.org
 * Fix AlertList panel generation
+* Add both upper and lower case `"time"` pattern for time_series column format in Table class
 * Drop testing of Python 2.7, it has been EOL'ed and CI was broken
   due to this.